option system implementation

+++ src/models/campeonatos.go
@@ -1,6 +1,6 @@
 package models
 
 type Campeonato struct {
-	ID     uint64 `json:"id,omitempty"`
+	ID     int `json:"id,omitempty"`
 	Titulo string `json:"titulo,omitempty"`
 }
